refactor: move pow2 operation trace to Bitwise processor
